+++ preprocessy/pipelines/_base.py
@@ -1,4 +1,5 @@
 import warnings
+from copy import deepcopy
 
 import stringcase
 from alive_progress import alive_bar
@@ -40,6 +41,8 @@ class BasePipeline:
     :param custom_reader: Custom function to read the data
     :type custom_reader: callable
 
+    .. versionchanged:: 1.0.4
+        Added ``__original_params`` and corresponding getter and setter to preserve original parameter dictionary.
     """
 
     def __init__(
@@ -51,7 +54,10 @@ class BasePipeline:
         params=None,
         custom_reader=None,
     ):
-        self.params = params
+        # self.params is now a copy of the original params dict that was passed
+        # .. versionadded:: 1.0.4
+        self.__original_params = params
+        self.params = deepcopy(self.__original_params)
         self.train_df_path = train_df_path
         self.test_df_path = test_df_path
         self.config_file = config_file
@@ -73,7 +79,8 @@ class BasePipeline:
         self.__validate_input()
 
         if self.config_file and not self.params:
-            self.params = read_config(self.config_file)
+            self.__original_params = read_config(self.config_file)
+            self.params = deepcopy(self.__original_params)
 
         if self.custom_reader is None:
             self.custom_reader = Reader().read_file
@@ -179,15 +186,17 @@ class BasePipeline:
         self.steps.insert(index, func)
         if params:
             for k, v in params.items():
-                if k in self.params:
+                if (k in self.params) or (k in self.__original_params):
                     raise ValueError(
                         f"Non unique parameter name. Param with name {k} already exists."
                     )
                 self.params[k] = v
+                self.__original_params[k] = v
 
     def add(self, func=None, params=None, **kwargs):
 
-        """Method to add another function to the pipeline after it has been constructed
+        """Method to add another function to the pipeline after it has been constructed. The
+        parameters of the newly added function are merged with ``self.params`` and ``self.__original_params``.
 
         :param func: The function to be added
         :type func: callable
@@ -211,8 +220,10 @@ class BasePipeline:
 
         :raises ArgumentsError: If no position is provided to insert the function into the pipeline
 
-        :raises ValueError: If ``params`` contains a key that already exists in ``self.params``
+        :raises ValueError: If ``params`` contains a key that already exists in ``self.params`` or ``self.__original_params``
 
+        .. versionchanged:: 1.0.4
+            Params are merged with both ``self.params`` and ``self.__original_params``.
         """
         if not callable(func):
             raise TypeError(
@@ -287,7 +298,9 @@ class BasePipeline:
         self.steps.remove(func)
 
     def save_config(self, file_path, config_drop_keys=None):
-        """Method to save the ``params`` to a ``JSON`` config file
+        """Method to save the ``params`` to a ``JSON`` config file. ``params`` is saved
+        instead of ``__original_params`` as some pipeline can generate or change values
+        of parameters and the most recent copy will be saved.
 
         :param file_path: Path where the config file must be created
         :type file_path: str
@@ -321,3 +334,35 @@ class BasePipeline:
         table.add_row(["Total Pipeline Stages", len(self.steps)])
         table.add_row(["Total Params", len(self.params.keys())])
         print(table)
+
+    def get_original_params(self):
+        """Returns the original parameter dictionary
+
+        :return: Original parameter dictionary
+        :rtype: dict
+
+        .. versionadded:: 1.0.4
+        """
+        return self.__original_params
+
+    def set_original_params(self, params):
+        """Set and overwrite the original parameters and ``self.params`` dictionaries
+        initialised in the constructor
+
+        :param params: A dictionary containing the parameters that are needed for configuring
+                        the pipeline.
+        :type params: dict
+        :raises ArgumentsError: ``params`` should not be a dictionary.
+
+        .. versionadded:: 1.0.4
+        """
+        if not isinstance(params, dict):
+            raise ArgumentsError("'params' should be a dictionary.")
+
+        self.__original_params = params
+
+        # Required for reader
+        self.__original_params["train_df_path"] = self.train_df_path
+        self.__original_params["test_df_path"] = self.test_df_path
+
+        self.params = deepcopy(self.__original_params)